refactor: Split LibraryAuthoringPage to components
